switches span header for caption and smaller improvements
